Add an atomic-check mutex for RTIC
This adds an AtomicCheck "mutex" which does a simple atomic check to
ensure that a bus collision does not occur and panics if one does. It is
then the user's responsibility to ensure that collisions do not occur
(e.g. by using RTIC to manage resources properly). This eliminates any
need to disable interrupts and supports high levels of concurrency.

Note that this is currently feature-gated behind `cortex-m` because the
implementation is only available for Cortex M platforms at the moment.