adc: add new proxy
Add support for ADC proxy. Note that OneShot trait describes a non-blocking
contract for ADC read operation. However access to shared ADC unit can not
be arbitrated in a completely non-blocking and concurrent way. Any reading
from a channel shall be completed before `shared-bus` can allow  next
reading from the same or another channel. So suggested implementation breaks
the non-blocking contract of the trait and just busy-spins until a sample
is returned.

Apparently this is kind of a hack. Possible better approach here would
be to change embedded-hal to have an explicitly blocking OneShot trait
instead and then make shared-bus work on top of that.

Meanwhile it worth mentioning that a lot of HALs do not implement ADC traits
in a non-blocking way. For instance, quite a few stm32 HALs actually block
until the reading is completed. The list of such HALs includes stm32f0xx-hal,
stm32f1xx-hal, stm32f4xx-hal to name a few.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>