For now, only pull in `atomic-polyfill` for `cortex-m` feature
It turns out the `AtomicCheckMutex` which was the reason for adding
`atomic-polyfill` is currently gated behing the `cortex-m` feature.
Thus we can gate the dependency behind this feature-flag to resolve
issues on other platforms in the meantime.

In the long run, `AtomicCheckMutex` should get its own feature-flag but
for now this is the easiest solution and a non-breaking change.

Ref: #31, #33